+++ android/anc/src/main/java/org/smartregister/fhircore/anc/data/patient/PatientRepository.kt
@@ -59,6 +59,7 @@ import org.smartregister.fhircore.anc.util.loadRegisterConfig
 import org.smartregister.fhircore.anc.util.loadRegisterConfigAnc
 import org.smartregister.fhircore.engine.data.domain.util.PaginationUtil
 import org.smartregister.fhircore.engine.data.domain.util.RegisterRepository
+import org.smartregister.fhircore.engine.data.local.DefaultRepository
 import org.smartregister.fhircore.engine.util.DispatcherProvider
 import org.smartregister.fhircore.engine.util.extension.due
 import org.smartregister.fhircore.engine.util.extension.extractAddress
@@ -89,7 +90,9 @@ constructor(
   private val ancOverviewConfig = context.loadRegisterConfigAnc(AncOverviewType.ANC_OVERVIEW_ID)
   private val vitalSignsConfig = context.loadRegisterConfigAnc(AncOverviewType.VITAL_SIGNS)
 
-  val resourceMapperExtended = ResourceMapperExtended(fhirEngine)
+  private val detailRepository = DefaultRepository(fhirEngine, dispatcherProvider)
+
+  val resourceMapperExtended = ResourceMapperExtended(detailRepository)
 
   override suspend fun loadData(
     query: String,
@@ -235,7 +238,7 @@ constructor(
         }
       }
     if (observations.isNotEmpty())
-      finalObservation = observations.sortedBy { it.effectiveDateTimeType.value }.first()
+      finalObservation = observations.sortedBy { it.effectiveDateTimeType.value }.last()
 
     return finalObservation
   }