+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/details/vitalsigns/VitalSignsDetailsViewModel.kt
@@ -21,9 +21,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import java.math.BigDecimal
 import javax.inject.Inject
 import kotlinx.coroutines.launch
+import org.hl7.fhir.r4.model.Observation
 import org.smartregister.fhircore.anc.data.model.EncounterItem
 import org.smartregister.fhircore.anc.data.model.PatientVitalItem
 import org.smartregister.fhircore.anc.data.patient.PatientRepository
@@ -51,30 +51,40 @@ constructor(val patientRepository: PatientRepository, var dispatcher: Dispatcher
   }
 
   fun fetchVitalSigns(patientId: String): LiveData<PatientVitalItem> {
-    val patientVitalOverviewItem = MutableLiveData<PatientVitalItem>()
+    val patientAncOverviewItem = MutableLiveData<PatientVitalItem>()
     val patientVitalItem = PatientVitalItem()
     viewModelScope.launch(dispatcher.io()) {
       val listObservationWeight =
         patientRepository.fetchVitalSigns(patientId = patientId, "body-weight")
       val listObservationHeight =
         patientRepository.fetchVitalSigns(patientId = patientId, "body-height")
+      val listObservationBps = patientRepository.fetchVitalSigns(patientId = patientId, "bp-s")
+      val listObservationBpds = patientRepository.fetchVitalSigns(patientId = patientId, "bp-d")
+      val listObservationPulseRate =
+        patientRepository.fetchVitalSigns(patientId = patientId, "pulse-rate")
+      val listObservationBg = patientRepository.fetchVitalSigns(patientId = patientId, "bg")
+      val listObservationspO2 = patientRepository.fetchVitalSigns(patientId = patientId, "spO2")
 
-      listObservationHeight
-        .valueQuantity
-        ?.value
-        ?.setScale(2, BigDecimal.ROUND_HALF_EVEN)
-        ?.toPlainString()
-        ?.let { patientVitalItem.height = it }
+      patientVitalItem.weight = observationValueOrDefault(listObservationWeight)
+      patientVitalItem.weightUnit = listObservationWeight.valueQuantity.unit ?: ""
 
-      listObservationWeight
-        .valueQuantity
-        ?.value
-        ?.setScale(2, BigDecimal.ROUND_HALF_EVEN)
-        ?.toPlainString()
-        ?.let { patientVitalItem.weight = it }
+      patientVitalItem.height = observationValueOrDefault(listObservationHeight)
+      patientVitalItem.heightUnit = listObservationHeight.valueQuantity.unit ?: ""
 
-      listObservationHeight.valueQuantity?.unit?.let { patientVitalItem.heightUnit = it }
-      listObservationWeight.valueQuantity?.unit?.let { patientVitalItem.weightUnit = it }
+      patientVitalItem.bps = observationValueOrDefault(listObservationBps)
+      patientVitalItem.bpsUnit = listObservationBps.valueQuantity.unit ?: ""
+
+      patientVitalItem.bpds = observationValueOrDefault(listObservationBpds)
+      patientVitalItem.bpdsUnit = listObservationBpds.valueQuantity.unit ?: ""
+
+      patientVitalItem.pulse = observationValueOrDefault(listObservationPulseRate)
+      patientVitalItem.pulseUnit = listObservationPulseRate.valueQuantity.unit ?: ""
+
+      patientVitalItem.bg = observationValueOrDefault(listObservationBg)
+      patientVitalItem.bgUnit = listObservationBg.valueQuantity.unit ?: ""
+
+      patientVitalItem.spO2 = observationValueOrDefault(listObservationspO2)
+      patientVitalItem.spO2Unit = listObservationspO2.valueQuantity.unit ?: ""
 
       if (patientVitalItem.height.isNotEmpty() &&
           patientVitalItem.weight.isNotEmpty() &&
@@ -101,8 +111,17 @@ constructor(val patientRepository: PatientRepository, var dispatcher: Dispatcher
           }
         }
       }
-      patientVitalOverviewItem.postValue(patientVitalItem)
+      patientAncOverviewItem.postValue(patientVitalItem)
     }
-    return patientVitalOverviewItem
+    return patientAncOverviewItem
+  }
+
+  private fun observationValueOrDefault(
+    observation: Observation,
+    defaultString: String = ""
+  ): String {
+    return if (observation.valueQuantity != null && observation.valueQuantity.value != null)
+      observation.valueQuantity.value.toPlainString()
+    else defaultString
   }
 }